backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.41.0
motor==3.7.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt<4.0.0
python-multipart==0.0.22
pydantic-settings==2.13.1
PyYAML==6.0.2
email-validator==2.2.0
